add bundles and changed script names
